Slår sammen ducks inn i innlogginsstatus for mer ryddighet
