+++ src/server/utils.ts
@@ -70,6 +70,7 @@ export const clientEnv = ({ req }: Props): Environment => {
                     LOGOUT_URL: req.query.logoutUrl as string,
                 }),
                 MASK_HOTJAR: req.query.maskHotjar !== 'false',
+                LOGOUT_WARNING: req.query.logoutWarning !== 'false',
             },
         }),
     };
